improved update.sh

+++ update.sh
@@ -1,3 +1,12 @@
 #!/bin/bash
-rm -rf /usr/bin/SLRHelper_0.1/
+if [ -f "/usr/bin/slrh" ];then
+	echo "slrh already installed : update the commande"
+else :
+	echo "slrh installation"
+	echo "#!/bin/bash" > /usr/bin/slrh
+	echo "python3 /usr/bin/SLRHelper_0.1/main.py $1" >> /usr/bin/slrh
+	chmod +x /usr/bin/slrh
+fi
+if [ -d "/usr/bin/SLRHelper_0.1" ]; then
+	rm -rf /usr/bin/SLRHelper_0.1/
 cp -r SLRHelper_0.1/ /usr/bin
